+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixed
 
--   [Patch] Prepend `$` to duration and easing tokens.
+-   [Patch] Prepend `$` to duration and easing SCSS tokens.
 
 ## 10.1.3 - 2019-12-10